+++ src/mesido/workflows/grow_workflow.py
@@ -151,7 +151,9 @@ class EndScenarioSizing(
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-        self.heat_network_settings["minimum_velocity"] = 0.0  # 0.001
+        # default setting to cater for ~ 10kW heat, DN800 pipe at dT = 40 degrees Celcuis
+        self.heat_network_settings["minimum_velocity"] = 1.0e-4
+
         self.heat_network_settings["maximum_velocity"] = 3.0
         self.heat_network_settings["head_loss_option"] = HeadLossOption.NO_HEADLOSS
 
@@ -250,7 +252,6 @@ class EndScenarioSizing(
         options = super().energy_system_options()
         options["maximum_temperature_der"] = np.inf
         options["heat_loss_disconnected_pipe"] = True
-        # options.update(self._override_hn_options)
         return options
 
     def path_goals(self):