+++ tests/test_updated_esdl_pre_process.py
@@ -7,6 +7,10 @@ from mesido.workflows import run_end_scenario_sizing
 
 import numpy as np
 
+from utils_test_scaling import create_problem_with_debug_info  # , problem_scaling_check
+
+from utils_tests import demand_matching_test, energy_conservation_test, heat_to_discharge_test
+
 
 class TestUpdatedESDL(TestCase):
     def test_updated_esdl(self):
@@ -33,12 +37,19 @@ class TestUpdatedESDL(TestCase):
             Path(examples.PoCTutorial.src.run_grow_tutorial.__file__).resolve().parent.parent
         )
 
+        optimscaling, logger, logs_list = create_problem_with_debug_info(
+            EndScenarioSizingStagedHighs
+        )
+
         problem = run_end_scenario_sizing(
-            EndScenarioSizingStagedHighs,
+            optimscaling,
             base_folder=base_folder,
             esdl_file_name="PoC Tutorial.esdl",
             esdl_parser=ESDLFileParser,
         )
+
+        # This code below is used to do manual check. Do not delete
+        # problem_scaling_check(logs_list, logger)
 
         # Save optimized esdl string
         optimized_esdl_string = problem.optimized_esdl_string
@@ -47,6 +58,10 @@ class TestUpdatedESDL(TestCase):
         )
         file.write(optimized_esdl_string)
         file.close()
+
+        demand_matching_test(problem, problem.extract_results())
+        energy_conservation_test(problem, problem.extract_results())
+        heat_to_discharge_test(problem, problem.extract_results())
 
         # Check the unique profile identification in profile_parser
         # Test that the correct demand profile is assigned to a demand as expected. Note that the
@@ -60,7 +75,7 @@ class TestUpdatedESDL(TestCase):
             # HeatingDemand_08fd
         )
         np.testing.assert_allclose(
-            724680.0,  # demand5_MW, multiplier 0.3
+            724680.0,  # demand5_MW, multiplier 0.33
             max(problem.get_timeseries("HeatingDemand_8fbe.target_heat_demand").values),
         )
         np.testing.assert_allclose(
@@ -68,19 +83,19 @@ class TestUpdatedESDL(TestCase):
             max(problem.get_timeseries("HeatingDemand_08fd.target_heat_demand").values),
         )
         np.testing.assert_allclose(
-            469709.62,  # demand4_MW, multiplier 0.75
+            475130.185,  # demand4_MW, multiplier 0.75
             np.average(problem.get_timeseries("HeatingDemand_b0ff.target_heat_demand").values),
         )
         np.testing.assert_allclose(
-            281825.77,  # demand5_MW, multiplier 0.3
+            285078.111,  # demand5_MW, multiplier 0.3
             np.average(problem.get_timeseries("HeatingDemand_8fbe.target_heat_demand").values),
         )
         np.testing.assert_allclose(
-            313139.75,  # demand4_MW, multiplier 0.5
+            316753.457,  # demand4_MW, multiplier 0.5
             np.average(problem.get_timeseries("HeatingDemand_08fd.target_heat_demand").values),
         )
 
-        # Checkk that the correct multiplier value was used
+        # Check that the correct multiplier value was used
         # Compare 2 max values where the same profile was used but with different multiplier values
         # HeatingDemand_08fd: multiplier 0.5
         # HeatingDemand_b0ff: multplier 0.75