+++ CHANGELOG.md
@@ -13,7 +13,8 @@
 - Bugfix: Scaling fix on ATES temperature variable when temperature modelling not used.
  
 ## Fixed
-- xxx
+- bug fix in grow_workflow (heat networks), changed minimum velocity from 0 to default value
+- addition to tests checking the minimum velocity setting in a workflow
 
 
 # [0.1.7] - 2024-08-23